Очистил роутер
