+++ package.json
@@ -21,5 +21,8 @@
   },
   "engines": {
     "node": "10"
+  },
+  "dependencies": {
+    "start": "^5.1.0"
   }
 }
